Remove makeRealRes from DispatchLogic
This becomes critical after the Doobie 0.9.0 upgrade because the
LiftDispatcher.makeResponse needs to execute on the same thread that the
request came in on. This is because ol' fashioned Lift uses thread-local
variables for its cookies and headers. After the Doobie upgrade,
dispatch logic often ends up on the EC or blocker[IO] thread pools.