Hide dead tag defaults when HideDead
This restores the original behaviour as it was before the previous
commit. The reason this is desirable is that we hide N/A tags too;
should keep it consistent. Plus I guess it's better to just have an
issue raised and all of your reqs blank in that field than to have the
dead tag appear all over the place.